+++ src/components/icons.js
@@ -1,0 +1,34 @@
+import CogIcon from '@rsuite/icons/legacy/Cog';
+import AngleLeftIcon from '@rsuite/icons/legacy/AngleLeft';
+import AngleRightIcon from '@rsuite/icons/legacy/AngleRight';
+import DashboardIcon from '@rsuite/icons/Dashboard';
+import GroupIcon from '@rsuite/icons/legacy/Group';
+import MagicIcon from '@rsuite/icons/legacy/Magic';
+import UserInfoIcon from '@rsuite/icons/UserInfo';
+import HelpOutlineIcon from '@rsuite/icons/HelpOutline';
+import NoticeIcon from '@rsuite/icons/Notice';
+import MemberIcon from '@rsuite/icons/Member';
+import UserChangeIcon from '@rsuite/icons/UserChange';
+import { IoMdContacts, IoMdPersonAdd, IoIosHome } from "react-icons/io";
+import { CgLogIn } from 'react-icons/cg'; 
+import InfoRoundIcon from '@rsuite/icons/InfoRound';
+
+
+export {
+    CogIcon,
+    AngleLeftIcon,
+    AngleRightIcon,
+    DashboardIcon,
+    GroupIcon,
+    MagicIcon,
+    UserInfoIcon,
+    HelpOutlineIcon,
+    NoticeIcon,
+    MemberIcon,
+    UserChangeIcon,
+    IoMdContacts, 
+    IoMdPersonAdd, 
+    IoIosHome,
+    CgLogIn,
+    InfoRoundIcon
+};
