+++ src/fregments/BaseBody.jsx
@@ -1,16 +1,18 @@
 import { Container, Header, Sidebar, Sidenav, Content, Navbar, Nav } from 'rsuite';
 import React from 'react';
-import CogIcon from '@rsuite/icons/legacy/Cog';
-import AngleLeftIcon from '@rsuite/icons/legacy/AngleLeft';
-import AngleRightIcon from '@rsuite/icons/legacy/AngleRight';
-import DashboardIcon from '@rsuite/icons/Dashboard';
-import GroupIcon from '@rsuite/icons/legacy/Group';
-import MagicIcon from '@rsuite/icons/legacy/Magic';
-import UserInfoIcon from '@rsuite/icons/UserInfo';
-import HelpOutlineIcon from '@rsuite/icons/HelpOutline';
-import NoticeIcon from '@rsuite/icons/Notice';
-import MemberIcon from '@rsuite/icons/Member';
-import UserChangeIcon from '@rsuite/icons/UserChange';
+import {
+    CogIcon,
+    AngleLeftIcon,
+    AngleRightIcon,
+    DashboardIcon,
+    GroupIcon,
+    MagicIcon,
+    UserInfoIcon,
+    HelpOutlineIcon,
+    NoticeIcon,
+    MemberIcon,
+    UserChangeIcon
+} from '@/components/icons.js';
 
 import { logo_image } from '@/assets/images'
 import { getCurrentPath } from '@/helpers/pathHelper';
