+++ src/fregments/BaseHeader.jsx
@@ -1,17 +1,20 @@
 import { Navbar, Nav } from 'rsuite';
 import React from 'react';
-import InfoRoundIcon from '@rsuite/icons/InfoRound';
-import { CgLogIn } from 'react-icons/cg'; 
-import HelpOutlineIcon from '@rsuite/icons/HelpOutline';
-import { IoMdContacts, IoMdPersonAdd, IoIosHome } from "react-icons/io";
-import DashboardIcon from '@rsuite/icons/Dashboard';
 
-import CogIcon from '@rsuite/icons/legacy/Cog';
 import { getAuthentication } from '@/helpers/authenHelpers';
 import { getCurrentPath } from '@/helpers/pathHelper';
 
 import { logo_image } from '@/assets/images'
-
+import {
+    CogIcon,
+    DashboardIcon,
+    HelpOutlineIcon,
+    IoMdContacts,
+    IoMdPersonAdd,
+    IoIosHome,
+    CgLogIn,
+    InfoRoundIcon
+} from '@/components/icons.js';
 
 const CustomNavbar = ({ onSelect, activeKey, ...props }) => {
     const auth = getAuthentication();
