+++ src/main.jsx
@@ -3,7 +3,7 @@ import App from '@/App.jsx'
 import '@/index.css'
 import "@fortawesome/fontawesome-free/css/all.min.css";
 import { Provider } from 'react-redux';
-import { Store } from '@/reduxs';
+import { Store } from '@/stores';
 
 
 ReactDOM.createRoot(document.getElementById('root')).render(
